requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-filter==23.1
psycopg2==2.9.6
django-redis==5.2.0
channels==4.0.0
asgiref==3.6.0
daphne==4.0.0
Pillow==9.4.0
drf-yasg==1.21.5
gunicorn==20.1.0
redis==5.0.0
celery==5.2.7
requests==2.28.2